refactor(assertions): share classifier prompt resolution, update edge baseline
Fixes the Style Check "architecture boundaries" failure on this branch.

The llama-guard handler had duplicated moderation.ts's chat-prompt extraction
(getActualPromptWithFallback + parseChatPrompt + last-user-message walk). That
duplication also duplicated its cross-layer imports, widening core -> providers
and core -> node past the ratchet in architecture/edge-baseline.json.

Extracting the logic into src/assertions/classifierPrompt.ts removes the
duplicate entirely: the core -> providers regression is gone, and core -> node
drops from +3 to +1. moderation.ts now imports the shared helper, so its
behavior is unchanged (its existing tests pass untouched).

The remaining baseline bump is unavoidable for any new assertion:

  core -> legacy-contracts  94 -> 97   AssertionParams/GradingResult/
                                       ProviderResponse type imports in the
                                       three new core files
  core -> node              90 -> 91   matchers/llamaGuard -> util/llamaGuard

Only those two edges are touched. The generated baseline also wanted to lower
providers -> legacy-runtime (325 -> 311) and providers -> node (301 -> 299),
but those reductions already exist on upstream/main independently of this
branch, so they are left alone rather than riding along with an unrelated
change; my +1 to providers -> node stays under the committed 301.
