+++ architecture/edge-baseline.json
@@ -13,9 +13,9 @@
   "cli -> redteam": 10,
   "cli -> view-server": 6,
   "core -> contracts": 1,
-  "core -> legacy-contracts": 94,
+  "core -> legacy-contracts": 97,
   "core -> legacy-runtime": 47,
-  "core -> node": 90,
+  "core -> node": 91,
   "core -> providers": 22,
   "core -> redteam": 8,
   "legacy-contracts -> contracts": 18,
